refactor: move GetFileExtension testing function to ./formats_test.go
