+++ base/component_test.go
@@ -1,13 +1,8 @@
 package base
 
 import (
-	"bufio"
 	_ "embed"
-	"encoding/base64"
 	"encoding/json"
-	"fmt"
-	"io"
-	"os"
 	"testing"
 
 	qt "github.com/frankban/quicktest"
@@ -52,25 +47,4 @@ func TestComponent_ListConnectorDefinitions(t *testing.T) {
 	err = json.Unmarshal(wantConnectorDefinitionJSON, &wantConnectorDefinitionStruct)
 	c.Assert(err, qt.IsNil)
 	c.Check(gotJSON, qt.JSONEquals, wantConnectorDefinitionStruct)
-}
-
-func TestUtil_GetFileExtension(t *testing.T) {
-	c := qt.New(t)
-
-	file, err := os.Open("./testdata/test_image.png")
-	c.Assert(err, qt.IsNil)
-	defer file.Close()
-	wantFileExtension := "png"
-
-	reader := bufio.NewReader(file)
-	content, err := io.ReadAll(reader)
-	c.Assert(err, qt.IsNil)
-
-	fileBase64 := base64.StdEncoding.EncodeToString(content)
-	fileBase64 = "data:image/png;base64," + fileBase64
-	fmt.Println(fileBase64)
-	gotFileExtension := GetBase64FileExtensionSlow(fileBase64)
-	c.Check(gotFileExtension, qt.Equals, wantFileExtension)
-	gotFileExtension = GetBase64FileExtensionFast(fileBase64)
-	c.Check(gotFileExtension, qt.Equals, wantFileExtension)
 }
